requirements.txt
fastapi
pyTelegramBotAPI
PyYAML
sdnotify
transliterate
uvicorn
